fix mesh write on 64-bit systems
git-svn-id: http://www.openmesh.org/svnrepo/OpenMesh/trunk@1280 fdac6126-5c0c-442c-9429-916003d36597

+++ src/OpenMesh/Tools/Decimater/ModProgMeshT.cc
@@ -160,9 +160,9 @@ write( const std::string& _ofname )
     {
       typename Mesh::ConstFaceVertexIter fv_it(Base::mesh(), *f_it);
       
-      IO::store( out, Base::mesh().property( idx_,   *fv_it ) );
-      IO::store( out, Base::mesh().property( idx_, *(++fv_it )) );
-      IO::store( out, Base::mesh().property( idx_, *(++fv_it )) );
+      IO::store( out, static_cast<unsigned int>(Base::mesh().property( idx_,   *fv_it )) );
+      IO::store( out, static_cast<unsigned int>(Base::mesh().property( idx_, *(++fv_it ))) );
+      IO::store( out, static_cast<unsigned int>(Base::mesh().property( idx_, *(++fv_it ))) );
     }
   }
   
@@ -172,11 +172,11 @@ write( const std::string& _ofname )
   { 
     // store v0.pos, v1.idx, vl.idx, vr.idx
     IO::store( out, vector_cast<Vec3f>(Base::mesh().point(r_it->v0)));
-    IO::store( out, Base::mesh().property( idx_, r_it->v1 ) );
+    IO::store(out, static_cast<unsigned int>(Base::mesh().property(idx_, r_it->v1)));
     IO::store( out, 
-	       r_it->vl.is_valid() ? Base::mesh().property(idx_, r_it->vl) : -1 );
+        r_it->vl.is_valid() ? static_cast<unsigned int>(Base::mesh().property(idx_, r_it->vl)) : -1);
     IO::store( out, 
-	       r_it->vr.is_valid() ? Base::mesh().property(idx_, r_it->vr) : -1 );
+        r_it->vr.is_valid() ? static_cast<unsigned int>(Base::mesh().property(idx_, r_it->vr)) : -1);
   }
 
   return true;
